Gbediv-85:  Accelerated peer review (#92)
- removed a bunch of scheduler forms that weren't in use & .a template we don't use
- replaced django-tinymce with downloading and using the package directly
- customized the new Tiny MCE 5 with a similar set of settings to what we had
- tested that this actually fixes the core of #85 

A key thing here was that django now includes HTML5 "required" tag with does not work when you have a TinyMCE form that is instantiated with no data.  Changed all TinyMCE using forms to not publish this HTML5 tag, which does NOT change whether or not the feild is required in the back end.