GBE-544:  remove fields we don't use (#555)
* remove fields we don't use

and ease up on some that are optional to make manual testing easier

* so easy - searchable admin fields

shoud've done this long ago.  Doing it everywhere

* a collection of searchable fields for GBE

It was a totally subjective selection of areas we work in the most and where it's most helpful.  Also, now that it's around, it's easier to add when needed for debugging.

* more for gbe admin

* scheduling done

* finish coverage, pepdiff OK