Transations show status of cancelled transactions
Found a rather large bug in the process - if we reuse purchasers between transactions (we do that now), then this shows all of a purchasers purchses, it didn't properly filter conference in per user view.

Also added messaging for what the colors red & green mean, which meant a build out of the intro message for the user view (which previously didn't exist).