All tests pass + TODO
TODO - remove usermessage for volunteer eval delete.  It was buggy to begin w. and delete change made it impossible.

BIG DEAL - found a fix to the django welcome message as home page - found out how to mock a django CMS home page so that our redirects and messages look right in test.